# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WEMU API Documentation",
        "description": "Swagger documentation for all APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/docs",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/artist/details/{id}": {
            "get": {
                "tags": [
                    "Artist"
                ],
                "summary": "Get artist details",
                "operationId": "2196fc66981f70a2d683f27cdfcdef89",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Artist details fetched"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}